deeplearning/requirements.txt
albumentations==1.1.0
deeplearning==0.0.1
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.1.1
responses==0.18.0
scikit_learn==1.1.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.62.3
opencv_python_headless==4.5.5.64
